-- Security / SessionHandler-- * Added CSRF protection - the core function now generates a random CSRF token for the session (called "CSRF_token"). The token is passed to the Front End, and is consumed (and regenerated) by various API endpoints on action.   Added it to most APIs, and added the generation and helper scripts to coreInit and sessionHandler.
* Added a timingManager utility class. Using it to add timing protection, currently only at the User API endpoint for some procedures.

-- Front End --
Made initPage more modular - now instead of reloading on specific events, has various hooks to insert callbacks on those events.
For example, now when you successfully log in, the default callback is a function that reloads the page (instead of the previously hardcoded boolean that either reloads the page or does nothing).